Fixed some issues

+++ server-awb-middleoffice/src/main/java/com/example/springsecuritypfe/model/CourbeBDT.java
@@ -30,7 +30,7 @@ import lombok.Data;
 	    private String dateCourbe;
 	    
 	    @Column(name="maturite")
-	    private Long maturite;  
+	    private Integer maturite;  
 
 	    @Column(name="tmp")
 	    private Double tmp;
