+++ server-awb-middleoffice/src/main/java/com/example/springsecuritypfe/service/CourbeLTService.java
@@ -15,7 +15,7 @@ public interface CourbeLTService {
     
     List<CourbeLT> findByDate(String date) throws ParseException;
     
-    List<CourbeLT> findByMaturite(Long maturite);
+    List<CourbeLT> findByMaturite(Integer maturite);
     
     List<CourbeLT> generateCourbeLT(String date) throws ParseException;
